w4BM v2.3.0
Fairly large reworks of series summaries which includes the addition of the series_works_summaries var (which only does anything when bookmarking a series) that can be used in workInfo to add the summaries of the works in the series to the series bookmark.
Added moment.js as a library in case anyone wants to use it to set their own date var.