CrWFC 2.2.1
Make sure page num is not captured when getting the normalised thread url. Move regexs and variables into the functions they are used in. Get auth name and url from the thread header area which is on every page of the thread as opposed to the first message of the thread